[improve] - menambahkan auditingtable

+++ src/main/java/com/bone/bone/repository/UserRepository.java
@@ -7,16 +7,17 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends DocoRepository<User, Long> {
+public interface UserRepository extends DocoRepository<User, Integer> {
     Optional<User> findByEmail(String email);
-
+    
     Optional<User> findByUsernameOrEmail(String username, String email);
 
-    List<User> findByIdIn(List<Long> userIds);
+    List<User> findByIdIn(List<Integer> userIds);
 
     Optional<User> findByUsername(String username);
 
     Boolean existsByUsername(String username);
 
     Boolean existsByEmail(String email);
+
 }
